(working) draw figure

+++ OS/Kernel/include/systemInfo.h
@@ -1,6 +1,6 @@
 #ifndef _SYSTEMINFO_H_
 #define _SYSTEMINFO_H_
 
-void cpuTemp(double * res);
+void cpuTemp(int * res);
 void localtime(char * buffer);
 #endif
